requirements.txt
aniso8601==9.0.1
attrs==23.2.0
backports.zoneinfo==0.2.1
blinker==1.8.2
cffi==1.16.0
click==8.1.7
contourpy==1.1.1
CoreNLG==2.2.0
cryptography==3.4.8
cycler==0.10.0
Flask==2.2.5
flask-restx==1.3.0
Flask-SQLAlchemy==2.5.1
fonttools==4.39.0
greenlet==1.1.0
gunicorn==20.1.0
importlib-metadata==4.0.1
importlib-resources==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==3.2.0
jwt==1.2.0
kiwisolver==1.3.1
lxml==4.9.3
MarkupSafe==2.1.1
matplotlib==3.6.3
mlxtend==0.19.0
numpy==1.19.5
packaging==21.3
pandas==1.2.5
Pillow==8.3.2
pkgutil_resolve_name==1.3.10
psycopg==3.0.8
psycopg2-binary==2.9.1
pycparser==2.20
pyparsing==2.4.7
pyrsistent==0.20.0
python-dateutil==2.8.1
python-dotenv==0.20.0
pytz==2021.1
referencing==0.28.0
rpds-py==0.18.1
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
SQLAlchemy==1.4.44
threadpoolctl==2.1.0
typing-extensions==3.10.0.2
tzdata==2021.1
Werkzeug==2.2.3
zipp==3.6.0
